+++ apps/coursequizzer/tests/unit/shims/app-state.ts
@@ -1,0 +1,9 @@
+export const page: {
+  params: {
+    courseId: string;
+  };
+} = {
+  params: {
+    courseId: '',
+  },
+};